+++ src/UglyToad.PdfPig/Graphics/Operations/PathConstruction/AppendRectangle.cs
@@ -1,7 +1,6 @@
 ﻿namespace UglyToad.PdfPig.Graphics.Operations.PathConstruction
 {
     using System.IO;
-    using PdfPig.Core;
 
     /// <inheritdoc />
     /// <remarks>
@@ -20,22 +19,22 @@
         /// <summary>
         /// The x coordinate of the lower left corner.
         /// </summary>
-        public decimal LowerLeftX { get; }
+        public double LowerLeftX { get; }
 
         /// <summary>
         /// The y coordinate of the lower left corner.
         /// </summary>
-        public decimal LowerLeftY { get; }
+        public double LowerLeftY { get; }
 
         /// <summary>
         /// The width of the rectangle.
         /// </summary>
-        public decimal Width { get; }
+        public double Width { get; }
 
         /// <summary>
         /// The height of the rectangle.
         /// </summary>
-        public decimal Height { get; }
+        public double Height { get; }
 
         /// <summary>
         /// Create a new <see cref="AppendRectangle"/>.
@@ -44,7 +43,7 @@
         /// <param name="y">The y coordinate of the lower left corner.</param>
         /// <param name="width">The width of the rectangle.</param>
         /// <param name="height">The height of the rectangle.</param>
-        public AppendRectangle(decimal x, decimal y, decimal width, decimal height)
+        public AppendRectangle(double x, double y, double width, double height)
         {
             LowerLeftX = x;
             LowerLeftY = y;
@@ -56,19 +55,19 @@
         /// <inheritdoc />
         public void Run(IOperationContext operationContext)
         {
-            operationContext.Rectangle((double)LowerLeftX, (double)LowerLeftY, (double)Width, (double)Height);
+            operationContext.Rectangle(LowerLeftX, LowerLeftY, Width, Height);
         }
 
         /// <inheritdoc />
         public void Write(Stream stream)
         {
-            stream.WriteDecimal(LowerLeftX);
+            stream.WriteDouble(LowerLeftX);
             stream.WriteWhiteSpace();
-            stream.WriteDecimal(LowerLeftY);
+            stream.WriteDouble(LowerLeftY);
             stream.WriteWhiteSpace();
-            stream.WriteDecimal(Width);
+            stream.WriteDouble(Width);
             stream.WriteWhiteSpace();
-            stream.WriteDecimal(Height);
+            stream.WriteDouble(Height);
             stream.WriteWhiteSpace();
             stream.WriteText(Symbol);
             stream.WriteNewLine();
